Database schema data type fix

+++ ui/scripts/2.2/create_obmdb_2.2.mysql.sql
@@ -464,7 +464,7 @@ CREATE TABLE `Contact` (
   `contact_comment`             text,
   `contact_comment2`            text,
   `contact_comment3`            text,
-  `contact_collected`           int(1) default '0',
+  `contact_collected`           tinyint(1) default '0',
   `contact_origin`              varchar(255) NOT NULL,
   PRIMARY KEY  (`contact_id`),
   KEY `contact_domain_id_domain_id_fkey` (`contact_domain_id`),
